Move implementation of ank get state into own file
Issue-Id: #229
