+++ ank/src/cli_commands.rs
@@ -12,18 +12,18 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{
-    collections::{HashMap, HashSet},
-    fmt::{self, Display},
-    time::Duration,
-};
-mod apply_manifests;
+use std::{collections::HashSet, time::Duration};
 pub mod server_connection;
 mod wait_list;
 use tokio::time::interval;
 use wait_list::WaitList;
 mod get_workload_table_display;
-use get_workload_table_display::{GetWorkloadTableDisplay, GetWorkloadTableDisplayWithSpinner};
+use get_workload_table_display::GetWorkloadTableDisplay;
+mod wait_list_display;
+
+// CLI commands implemented in another files
+mod apply_manifests;
+mod get_state;
 
 #[cfg(not(test))]
 async fn read_file_to_string(file: String) -> std::io::Result<String> {
@@ -40,162 +40,13 @@ use common::{
 };
 
 use tabled::{settings::Style, Table};
+use wait_list_display::WaitListDisplay;
 
 #[cfg_attr(test, mockall_double::double)]
 use self::server_connection::ServerConnection;
-use self::wait_list::WaitListDisplayTrait;
 use crate::{
-    cli::OutputFormat, cli_commands::wait_list::ParsedUpdateStateSuccess, cli_error::CliError,
-    output, output_and_error, output_debug,
+    cli_commands::wait_list::ParsedUpdateStateSuccess, cli_error::CliError, output, output_debug,
 };
-
-const SPINNER_SYMBOLS: [&str; 4] = ["|", "/", "-", "\\"];
-pub(crate) const COMPLETED_SYMBOL: &str = " ";
-
-fn generate_compact_state_output(
-    state: &CompleteState,
-    object_field_mask: Vec<String>,
-    output_format: OutputFormat,
-) -> Result<String, CliError> {
-    let convert_to_output = |map: serde_yaml::Value| -> Result<String, CliError> {
-        match output_format {
-            // [impl -> swdd~cli-shall-support-desired-state-yaml~1]
-            OutputFormat::Yaml => Ok(serde_yaml::to_string(&map)?),
-            // [impl -> swdd~cli-shall-support-desired-state-json~1]
-            OutputFormat::Json => Ok(serde_json::to_string_pretty(&map)?),
-        }
-    };
-
-    let deserialized_state: serde_yaml::Value = serde_yaml::to_value(state)?;
-
-    if object_field_mask.is_empty() {
-        return convert_to_output(deserialized_state);
-    }
-
-    let mut compact_state = serde_yaml::Value::Mapping(Default::default());
-    for mask in object_field_mask {
-        let splitted_masks: Vec<&str> = mask.split('.').collect();
-        if let Some(filtered_mapping) = get_filtered_value(&deserialized_state, &splitted_masks) {
-            update_compact_state(
-                &mut compact_state,
-                &splitted_masks,
-                filtered_mapping.to_owned(),
-            );
-        }
-    }
-
-    convert_to_output(compact_state)
-}
-
-fn get_filtered_value<'a>(
-    map: &'a serde_yaml::Value,
-    mask: &[&str],
-) -> Option<&'a serde_yaml::Value> {
-    mask.iter()
-        .try_fold(map, |current_level, mask_part| current_level.get(mask_part))
-}
-
-fn update_compact_state(
-    new_compact_state: &mut serde_yaml::Value,
-    mask: &[&str],
-    new_mapping: serde_yaml::Value,
-) -> Option<()> {
-    if mask.is_empty() {
-        return Some(());
-    }
-
-    let mut current_level = new_compact_state;
-
-    for mask_part in mask {
-        if current_level.get(mask_part).is_some() {
-            current_level = current_level.get_mut(mask_part)?;
-            continue;
-        }
-
-        if let serde_yaml::Value::Mapping(current_mapping) = current_level {
-            current_mapping.insert(
-                (*mask_part).into(),
-                serde_yaml::Value::Mapping(Default::default()),
-            );
-
-            current_level = current_mapping.get_mut(mask_part)?;
-        } else {
-            return None;
-        }
-    }
-
-    *current_level = new_mapping;
-    Some(())
-}
-
-struct WaitListDisplay {
-    data: HashMap<WorkloadInstanceName, GetWorkloadTableDisplay>,
-    not_completed: HashSet<WorkloadInstanceName>,
-    spinner: Spinner,
-}
-
-impl Display for WaitListDisplay {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let current_spinner = self.spinner.to_string();
-        let mut data: Vec<_> = self
-            .data
-            .iter()
-            .map(|(workload_name, table_entry)| {
-                let update_state_symbol = if self.not_completed.contains(workload_name) {
-                    &current_spinner
-                } else {
-                    COMPLETED_SYMBOL
-                };
-                GetWorkloadTableDisplayWithSpinner {
-                    data: table_entry,
-                    spinner: update_state_symbol,
-                }
-            })
-            .collect();
-        data.sort_by_key(|x| &x.data.name);
-
-        // [impl->swdd~cli-shall-present-workloads-as-table~1]
-        write!(
-            f,
-            "{}",
-            Table::new(data).with(tabled::settings::Style::blank())
-        )
-    }
-}
-
-impl WaitListDisplayTrait for WaitListDisplay {
-    fn update(&mut self, workload_state: &common::objects::WorkloadState) {
-        if let Some(entry) = self.data.get_mut(&workload_state.instance_name) {
-            entry.execution_state = workload_state.execution_state.state.to_string();
-            entry.additional_info = workload_state.execution_state.additional_info.clone();
-        }
-    }
-
-    fn set_complete(&mut self, workload: &WorkloadInstanceName) {
-        self.not_completed.remove(workload);
-    }
-
-    fn step_spinner(&mut self) {
-        self.spinner.step();
-    }
-}
-
-#[derive(Default)]
-struct Spinner {
-    pos: usize,
-}
-
-impl Spinner {
-    pub fn step(&mut self) {
-        self.pos = (self.pos + 1) % SPINNER_SYMBOLS.len();
-    }
-}
-
-impl Display for Spinner {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", SPINNER_SYMBOLS[self.pos])
-    }
-}
 
 pub struct CliCommands {
     // Left here for the future use.
@@ -220,34 +71,6 @@ impl CliCommands {
 
     pub async fn shut_down(self) {
         self.server_connection.shut_down().await
-    }
-
-    pub async fn get_state(
-        &mut self,
-        object_field_mask: Vec<String>,
-        output_format: OutputFormat,
-    ) -> Result<String, CliError> {
-        output_debug!(
-            "Got: object_field_mask={:?} output_format={:?}",
-            object_field_mask,
-            output_format
-        );
-
-        let res_complete_state = self
-            .server_connection
-            .get_complete_state(&object_field_mask)
-            .await?;
-        // [impl->swdd~cli-returns-api-version-with-desired-state~1]
-        // [impl->swdd~cli-returns-api-version-with-startup-state~1]
-        // [impl->swdd~cli-returns-compact-state-object-when-object-field-mask-provided~1]
-        match generate_compact_state_output(&res_complete_state, object_field_mask, output_format) {
-            Ok(res) => Ok(res),
-            Err(err) => {
-                output_and_error!(
-                    "Error occurred during processing response from server.\nError: {err}"
-                );
-            }
-        }
     }
 
     fn add_default_workload_spec_per_update_mask(
@@ -585,7 +408,7 @@ mod tests {
             StoredWorkloadSpec, Tag, WorkloadState,
         },
         state_manipulation::{Object, Path},
-        test_utils::{self, generate_test_complete_state},
+        test_utils::{self},
         to_server_interface::ToServerReceiver,
     };
     use mockall::predicate::eq;
@@ -597,11 +420,8 @@ mod tests {
         handle_agent_overwrite, parse_manifest, update_request_obj, InputSourcePair,
     };
     use crate::{
-        cli::{ApplyArgs, OutputFormat},
-        cli_commands::{
-            generate_compact_state_output, get_filtered_value,
-            server_connection::MockServerConnection, update_compact_state, GetWorkloadTableDisplay,
-        },
+        cli::ApplyArgs,
+        cli_commands::{server_connection::MockServerConnection, GetWorkloadTableDisplay},
     };
     use serde_yaml::Value;
     use std::io::Read;
@@ -609,20 +429,6 @@ mod tests {
     use super::CliCommands;
 
     const RESPONSE_TIMEOUT_MS: u64 = 3000;
-
-    const EXAMPLE_STATE_INPUT: &str = r#"{
-        "desiredState": {
-            "workloads": {
-                "nginx": {
-                    "restartPolicy": ALWAYS,
-                    "agent": "agent_A"
-                },
-                "hello1": {
-                    "agent": "agent_B"
-                }
-            }
-        }
-    }"#;
 
     mockall::lazy_static! {
         pub static ref FAKE_READ_TO_STRING_MOCK_RESULT_LIST: tokio::sync::Mutex<std::collections::VecDeque<io::Result<String>>>  =
@@ -1030,249 +836,6 @@ mod tests {
         assert!(delete_result.is_ok());
     }
 
-    // [utest -> swdd~cli-returns-desired-state-from-server~1]
-    // [utest -> swdd~cli-shall-support-desired-state-yaml~1]
-    // [utest->swdd~cli-blocks-until-ankaios-server-responds-get-desired-state~1]
-    // [utest->swdd~cli-provides-get-desired-state~1]
-    #[tokio::test]
-    async fn utest_get_state_complete_desired_state_yaml() {
-        let test_data = test_utils::generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "runtime".to_string(),
-            ),
-        ]);
-
-        let mut mock_server_connection = MockServerConnection::default();
-        let test_data_clone = test_data.clone();
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec![]))
-            .return_once(|_| Ok(Box::new(test_data_clone)));
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(vec![], crate::cli::OutputFormat::Yaml)
-            .await
-            .unwrap();
-        let expected_text = serde_yaml::to_string(&test_data).unwrap();
-        assert_eq!(cmd_text, expected_text);
-    }
-
-    // [utest -> swdd~cli-shall-support-desired-state-json~1]
-    #[tokio::test]
-    async fn utest_get_state_complete_desired_state_json() {
-        let test_data = test_utils::generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "runtime".to_string(),
-            ),
-        ]);
-
-        let mut mock_server_connection = MockServerConnection::default();
-        let cloned_test_data = test_data.clone();
-        mock_server_connection
-            .expect_get_complete_state()
-            .return_once(|_| Ok(Box::new(cloned_test_data)));
-
-        let mut cmd = CliCommands {
-            _response_timeout_ms: 0,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(vec![], crate::cli::OutputFormat::Json)
-            .await
-            .unwrap();
-
-        let expected_text = serde_json::to_string_pretty(&test_data).unwrap();
-        assert_eq!(cmd_text, expected_text);
-    }
-
-    // [utest->swdd~cli-returns-api-version-with-startup-state~1]
-    #[tokio::test]
-    async fn utest_get_state_startup_state() {
-        let test_data = CompleteState::default();
-
-        let mut mock_server_connection = MockServerConnection::default();
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec!["startupState".to_owned()]))
-            .return_once(|_| Ok(Box::new(test_data)));
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(
-                vec!["startupState".to_owned()],
-                crate::cli::OutputFormat::Yaml,
-            )
-            .await
-            .unwrap();
-
-        let expected_single_field_result_text =
-            "startupState:\n  apiVersion: v0.1\n  workloads: {}\n";
-
-        assert_eq!(cmd_text, expected_single_field_result_text);
-    }
-
-    // [utest -> swdd~cli-returns-desired-state-from-server~1]
-    // [utest->swdd~cli-returns-api-version-with-desired-state~1]
-    #[tokio::test]
-    async fn utest_get_state_single_field_of_desired_state() {
-        let test_data = test_utils::generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "runtime".to_string(),
-            ),
-        ]);
-
-        let mut mock_server_connection = MockServerConnection::default();
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec!["desiredState.workloads.name3.runtime".into()]))
-            .return_once(|_| Ok(Box::new(test_data)));
-
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(
-                vec!["desiredState.workloads.name3.runtime".to_owned()],
-                crate::cli::OutputFormat::Yaml,
-            )
-            .await
-            .unwrap();
-
-        let expected_single_field_result_text =
-            "desiredState:\n  workloads:\n    name3:\n      runtime: runtime\n";
-
-        assert_eq!(cmd_text, expected_single_field_result_text);
-    }
-
-    // [utest->swdd~cli-provides-object-field-mask-arg-to-get-partial-desired-state~1]
-    // [utest->swdd~cli-returns-compact-state-object-when-object-field-mask-provided~1]
-    // [utest->swdd~cli-returns-api-version-with-desired-state~1]
-    #[tokio::test]
-    async fn utest_get_state_multiple_fields_of_desired_state() {
-        let test_data = test_utils::generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "runtime".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "runtime".to_string(),
-            ),
-        ]);
-
-        let mut mock_server_connection = MockServerConnection::default();
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec![
-                "desiredState.workloads.name1.runtime".into(),
-                "desiredState.workloads.name2.runtime".into(),
-            ]))
-            .return_once(|_| Ok(Box::new(test_data)));
-
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(
-                vec![
-                    "desiredState.workloads.name1.runtime".to_owned(),
-                    "desiredState.workloads.name2.runtime".to_owned(),
-                ],
-                crate::cli::OutputFormat::Yaml,
-            )
-            .await
-            .unwrap();
-        assert!(matches!(cmd_text,
-            txt if txt == *"desiredState:\n  workloads:\n    name1:\n      runtime: runtime\n    name2:\n      runtime: runtime\n" ||
-            txt == *"desiredState:\n  workloads:\n    name2:\n      runtime: runtime\n    name1:\n      runtime: runtime\n"));
-    }
-
-    #[tokio::test]
-    async fn utest_get_state_single_field_without_api_version() {
-        let test_data = test_utils::generate_test_complete_state(Vec::new());
-
-        let mut mock_server_connection = MockServerConnection::default();
-        mock_server_connection
-            .expect_get_complete_state()
-            .with(eq(vec!["workloadStates".to_owned()]))
-            .return_once(|_| Ok(Box::new(test_data)));
-        let mut cmd = CliCommands {
-            _response_timeout_ms: RESPONSE_TIMEOUT_MS,
-            no_wait: false,
-            server_connection: mock_server_connection,
-        };
-
-        let cmd_text = cmd
-            .get_state(
-                vec!["workloadStates".to_owned()],
-                crate::cli::OutputFormat::Yaml,
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(cmd_text, "workloadStates: []\n");
-    }
-
     // [utest->swdd~cli-provides-run-workload~1]
     // [utest->swdd~cli-blocks-until-ankaios-server-responds-run-workload~2]
     #[tokio::test]
@@ -1357,304 +920,6 @@ mod tests {
             )
             .await;
         assert!(run_workload_result.is_ok());
-    }
-
-    #[test]
-    fn utest_generate_compact_state_output_empty_filter_masks() {
-        let input_state = generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "podman".to_string(),
-            ),
-        ]);
-
-        let cli_output =
-            generate_compact_state_output(&input_state, vec![], OutputFormat::Yaml).unwrap();
-
-        // state shall remain unchanged
-        assert_eq!(cli_output, serde_yaml::to_string(&input_state).unwrap());
-    }
-
-    #[test]
-    fn utest_generate_compact_state_output_single_filter_mask() {
-        let input_state = generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "podman".to_string(),
-            ),
-        ]);
-
-        let expected_state = r#"{
-            "desiredState": {
-                "workloads": {
-                  "name1": {
-                    "agent": "agent_A",
-                    "tags": [
-                        {
-                        "key": "key",
-                        "value": "value"
-                        }
-                    ],
-                    "dependencies": {
-                        "workload A": "ADD_COND_RUNNING",
-                        "workload C": "ADD_COND_SUCCEEDED"
-                    },
-                    "restartPolicy": "ALWAYS",
-                    "runtime": "podman",
-                    "runtimeConfig": "generalOptions: [\"--version\"]\ncommandOptions: [\"--network=host\"]\nimage: alpine:latest\ncommandArgs: [\"bash\"]\n"
-                    }
-                }
-            }
-        }"#;
-
-        let cli_output = generate_compact_state_output(
-            &input_state,
-            vec!["desiredState.workloads.name1".to_string()],
-            OutputFormat::Yaml,
-        )
-        .unwrap();
-
-        let expected_value: serde_yaml::Value = serde_yaml::from_str(expected_state).unwrap();
-
-        assert_eq!(cli_output, serde_yaml::to_string(&expected_value).unwrap());
-    }
-
-    #[test]
-    fn utest_generate_compact_state_output_multiple_filter_masks() {
-        let input_state = generate_test_complete_state(vec![
-            generate_test_workload_spec_with_param(
-                "agent_A".to_string(),
-                "name1".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name2".to_string(),
-                "podman".to_string(),
-            ),
-            generate_test_workload_spec_with_param(
-                "agent_B".to_string(),
-                "name3".to_string(),
-                "podman".to_string(),
-            ),
-        ]);
-
-        let expected_state = r#"{
-            "desiredState": {
-                "workloads": {
-                    "name1": {
-                        "agent": "agent_A",
-                        "tags": [
-                            {
-                            "key": "key",
-                            "value": "value"
-                            }
-                        ],
-                        "dependencies": {
-                            "workload A": "ADD_COND_RUNNING",
-                            "workload C": "ADD_COND_SUCCEEDED"
-                        },
-                        "restartPolicy": "ALWAYS",
-                        "runtime": "podman",
-                        "runtimeConfig": "generalOptions: [\"--version\"]\ncommandOptions: [\"--network=host\"]\nimage: alpine:latest\ncommandArgs: [\"bash\"]\n"
-                    },
-                    "name2": {
-                        "agent": "agent_B"
-                    }
-                }
-            }
-        }"#;
-
-        let cli_output = generate_compact_state_output(
-            &input_state,
-            vec![
-                "desiredState.workloads.name1".to_string(),
-                "desiredState.workloads.name2.agent".to_string(),
-            ],
-            OutputFormat::Yaml,
-        )
-        .unwrap();
-
-        let expected_value: serde_yaml::Value = serde_yaml::from_str(expected_state).unwrap();
-
-        assert_eq!(cli_output, serde_yaml::to_string(&expected_value).unwrap());
-    }
-
-    #[test]
-    fn utest_get_filtered_value_filter_key_with_mapping() {
-        let deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-        let result =
-            get_filtered_value(&deserialized_map, &["desiredState", "workloads", "nginx"]).unwrap();
-        assert_eq!(
-            result.get("restartPolicy").unwrap(),
-            &serde_yaml::Value::String("ALWAYS".into())
-        );
-    }
-
-    #[test]
-    fn utest_get_filtered_value_filter_key_without_mapping() {
-        let deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-        let result = get_filtered_value(
-            &deserialized_map,
-            &["desiredState", "workloads", "nginx", "agent"],
-        )
-        .unwrap();
-        let expected = serde_yaml::Value::String("agent_A".to_string());
-        assert_eq!(result, &expected);
-    }
-
-    #[test]
-    fn utest_get_filtered_value_empty_mask() {
-        let deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-        let result = get_filtered_value(&deserialized_map, &[]).unwrap();
-        assert!(result.get("desiredState").is_some());
-    }
-
-    #[test]
-    fn utest_get_filtered_value_not_existing_keys() {
-        let deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-
-        let result = get_filtered_value(
-            &deserialized_map,
-            &["desiredState", "workloads", "notExistingWorkload", "nginx"],
-        );
-        assert!(result.is_none());
-
-        let result = get_filtered_value(
-            &deserialized_map,
-            &[
-                "desiredState",
-                "workloads",
-                "notExistingWorkload",
-                "notExistingField",
-            ],
-        );
-        assert!(result.is_none());
-
-        let result = get_filtered_value(
-            &deserialized_map,
-            &[
-                "desiredState",
-                "workloads",
-                "nginx",
-                "agent",
-                "notExistingField",
-            ],
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn utest_update_compact_state_create_two_keys() {
-        let mut deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-
-        // update by inserting two new nested keys and a new empty mapping as value
-        update_compact_state(
-            &mut deserialized_map,
-            &[
-                "desiredState",
-                "workloads",
-                "createThisKey",
-                "createThisKey",
-            ],
-            serde_yaml::Value::Mapping(Default::default()),
-        );
-
-        assert!(deserialized_map
-            .get("desiredState")
-            .and_then(|next| next.get("workloads").and_then(|next| next
-                .get("createThisKey")
-                .and_then(|next| next.get("createThisKey"))))
-            .is_some());
-    }
-
-    #[test]
-    fn utest_update_compact_state_keep_value_of_existing_key() {
-        let mut deserialized_map: serde_yaml::Value =
-            serde_yaml::from_str(EXAMPLE_STATE_INPUT).unwrap();
-        // do not update value of existing key
-        update_compact_state(
-            &mut deserialized_map,
-            &[
-                "desiredState",
-                "workloads",
-                "nginx",
-                "restartPolicy",
-                "createThisKey",
-            ],
-            serde_yaml::Value::Mapping(Default::default()),
-        );
-
-        assert_eq!(
-            deserialized_map
-                .get("desiredState")
-                .and_then(|next| next
-                    .get("workloads")
-                    .and_then(|next| next.get("nginx").and_then(|next| next.get("restartPolicy"))))
-                .unwrap(),
-            &serde_yaml::Value::String("ALWAYS".into())
-        );
-    }
-
-    #[test]
-    fn utest_update_compact_state_insert_into_empty_map() {
-        // insert keys nested into empty map and add empty mapping as value
-        let mut empty_map = serde_yaml::Value::Mapping(Default::default());
-        update_compact_state(
-            &mut empty_map,
-            &["desiredState", "workloads", "nginx"],
-            serde_yaml::Value::Mapping(Default::default()),
-        );
-
-        assert!(empty_map
-            .get("desiredState")
-            .and_then(|next| next.get("workloads").and_then(|next| next.get("nginx")))
-            .is_some());
-    }
-
-    #[test]
-    fn utest_update_compact_state_do_not_update_on_empty_mask() {
-        let mut empty_map = serde_yaml::Value::Mapping(Default::default());
-        empty_map.as_mapping_mut().unwrap().insert(
-            "desiredState".into(),
-            serde_yaml::Value::Mapping(Default::default()),
-        );
-        let expected_map = empty_map.clone();
-
-        // do not update map if no masks are provided
-        update_compact_state(
-            &mut empty_map,
-            &[],
-            serde_yaml::Value::Mapping(Default::default()),
-        );
-        assert_eq!(empty_map, expected_map);
     }
 
     // [utest->swdd~cli-apply-supports-ankaios-manifest~1]
